minimal_app/django/requirements.txt
Django==4.0.2
psycopg2-binary==2.9.1
djangorestframework==3.13.1